refactor: mypageleftcard 리펙토링
