+++ src/components/pages/Mypage/MypageRentPage/index.tsx
@@ -1,69 +1,22 @@
-import { TRentInfo } from "@/components/atoms/Mypage/MypageRentSection";
 import { TRentContentInfo } from "@/components/molecules/Mypage/MypageRentList";
 import MypageLeftCard from "@/components/organisms/Mypage/MypageLeftCard";
 import MypageRentCard from "@/components/organisms/Mypage/MypageRentCard";
-import { loginInfo, rentHistories } from "@/recoil";
+import { rentHistories } from "@/recoil";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useRecoilValueLoadable } from "recoil";
 
 const MypageRentPage = () => {
-  const [userName, setUserName] = useState<string>("");
-  const [totalRentNum, setTotalRentNum] = useState<number>(0);
-  const [profileRentInfo, setProfileRentInfo] = useState<TRentInfo>({
-    umbrellaUuid: 0,
-    rentedAt: "",
-    rentedStore: "",
-    returnedDue: "",
-    returnAt: "",
-    returned: false,
-    refunded: false,
-  });
   const [rentList, setRentList] = useState<TRentContentInfo[]>([]);
   const umbrellaHistories = useRecoilValueLoadable(rentHistories);
-  const umbrellaNum = useRecoilValueLoadable(rentHistories);
-  const loginInfoValue = useRecoilValueLoadable(loginInfo);
 
   const navigate = useNavigate();
 
   useEffect(() => {
-    // left card의 유저 이름, 빌린 우산 수 가져오는 함수
-    const getUserInfo = async () => {
-      switch (loginInfoValue.state) {
-        case "hasValue":
-          setUserName(loginInfoValue.contents.name);
-          break;
-        case "loading":
-          <div>Loading...</div>;
-          break;
-        case "hasError":
-          alert("로그인 세션이 만료되었습니다. 다시 로그인 해주세요.");
-          navigate("/login");
-          break;
-      }
-      switch (umbrellaNum.state) {
-        case "hasValue":
-          setTotalRentNum(umbrellaNum.contents.length);
-          break;
-        case "loading":
-          <div>Loading...</div>;
-          break;
-        case "hasError":
-          alert("로그인 세션이 만료되었습니다. 다시 로그인 해주세요.");
-          navigate("/login");
-          break;
-      }
-    };
-    // left card의 대여중인 우산 가져오는 함수
-    const getCurrentUmbrellaId = async () => {
+    // rent list 가져오는 함수
+    const getRentList = async () => {
       switch (umbrellaHistories.state) {
         case "hasValue": {
-          const profileRent: TRentInfo = umbrellaHistories.contents.find(
-            (history: TRentInfo) => !history.returned
-          );
-
-          setProfileRentInfo({ ...profileRent });
-          // TODO:지금은 여기에 포함되어있지만 반드시 분리해줘야한다.
           setRentList(umbrellaHistories.contents);
           break;
         }
@@ -71,14 +24,11 @@ const MypageRentPage = () => {
           <div>Loading...</div>;
           break;
         case "hasError":
-          alert("로그인 세션이 만료되었습니다. 다시 로그인 해주세요.");
-          navigate("/login");
           break;
       }
     };
-    getUserInfo();
-    getCurrentUmbrellaId();
-  }, [loginInfoValue, umbrellaNum, umbrellaHistories, navigate]);
+    getRentList();
+  }, [umbrellaHistories, navigate]);
 
   return (
     <div className="flex justify-center w-[1280px] mt-24 px-40">
@@ -86,12 +36,7 @@ const MypageRentPage = () => {
         <div className="text-black text-24 font-semibold leading-32 mb-32">MYPAGE</div>
         <div className="flex">
           <div className="mr-32">
-            <MypageLeftCard
-              userName={userName}
-              totalRentNum={totalRentNum}
-              profileInfo={profileRentInfo}
-              isReturned={profileRentInfo.returned}
-            />
+            <MypageLeftCard />
           </div>
           <MypageRentCard rentList={rentList} />
           {/*//TODO: 이부분의 components만 상황에 따라 바뀌도록 할 수 있으면 더 깔끔할텐데..? */}
